+++ word_search_generator/__init__.py
@@ -8,7 +8,7 @@
     :license: MIT, see LICENSE for more details.
 """
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 
 from . import export
 from . import generate