+++ CHANGES.md
@@ -24,3 +24,20 @@ if args.export and args.path is None:
     if args.export and args.path is None:
         args.path = pathlib.Path.cwd()
 ```
+
+**1.0.2** (2021-08-17)
+
+-   Added some cli testing
+-   Updated cli -p, --path flag to more standard -o, --output flag
+-   Updated cli to accept stdin of words instead of requiring -w or --words flag
+
+```bash
+# instead of using -w just list word as arguments
+$ words-search some sample words to include
+# all other flags work the same
+$ words-search some sample words to include -l 3 -o pdf
+# you can also pipe directly to word-search
+$ cat some-sample-words.txt | word-search
+# pipe a bunch of files for export
+$ for f in tests/word*.txt; do word-search "$(cat $f)" -e pdf -o $f.pdf; done
+```